backend/src/scripts/requirements.txt
transformers
torch
Pillow
numpy
scipy